adds the ability to process a cucumber rerun file
Cucumber can be instructed to output a list of any failed tests in a run.
It is then possible to run only those tests by specifying the file that
was written by cucumber and passing it into the next test run.
Cucumber already supports this ability natively, adding it as an option
to parallel_calabash wrapper.